add emit select repl

+++ dashboard/utils/orders/clean_orders.py
@@ -5,9 +5,11 @@ datastore_client = datastore.Client()
 def delete_all_orders(kind="orders"):
     query = datastore_client.query(kind=kind)
     all_entities = list(query.fetch())
-    for entitiy in all_entities:
-        key = entitiy.key
-        datastore_client.delete(key)
+    resp = input('enter <y> to delete all')
+    if resp == 'y':
+        for entitiy in all_entities:
+            key = entitiy.key
+            datastore_client.delete(key)
 
 
 delete_all_orders()
